+++ tests/apache-php/run_image_tests.sh
@@ -7,7 +7,7 @@ IMAGE_TAG=$1
 
 CURRENT_DIR=$(dirname $(readlink -f $0))
 
-docker run -d --name akeneo-apache-php-${IMAGE_TAG} -u docker -v ${CURRENT_DIR}:${CURRENT_DIR} akeneo/apache-php:${IMAGE_TAG}
+docker run -d --name akeneo-apache-php-${IMAGE_TAG} -u 1000:1000 -v ${CURRENT_DIR}:${CURRENT_DIR} akeneo/apache-php:${IMAGE_TAG}
 
 for TEST in $(ls -1 ${CURRENT_DIR}/common/*.sh 2> /dev/null || true); do
     docker exec akeneo-apache-php-${IMAGE_TAG} bash ${TEST} || DID_FAIL=1